refactor(core): unify cross-platform variant merge rules in platform-variants
Split pages-json.ts (1174 lines) into four files:

- platform-variants: the shared variant table core. Five merge rules
  (exclude current platform, stack platforms on equal values, keep
  hand-written #ifndef untouched, stable sort by platformStr, wrap
  unless the union is covered) now live in one place.
- jsonc-scan: raw-text scanner that recovers duplicate keys
  comment-json drops when parsing.
- tab-bar: object-side adapter for tabBar look-and-feel props plus
  hand-rendered output lines.
- pages-json: array-side adapter, merge orchestration, home page
  ordering, file lock and atomic write.

Both adapters feed the shared core (upsertCurrentVariant, sortVariants,
union helpers) instead of keeping two parallel copies of the rules.
