+++ packages/core/src/pages-json.ts
@@ -1,17 +1,32 @@
 import type { Pages, PagesConfig, SubPackage, SubPackages, TabBar } from '@uni-helper/uni-pages-types'
 import type { CommentLineToken, CommentObject, CommentSymbol, CommentToken } from 'comment-json'
+import type { PlatformVariant } from './platform-variants'
 import type { ExcludeIndexSignature, InternalPageItem, InternalPages } from './types'
 import fs from 'node:fs'
 import { parse as cjParse, stringify as cjStringify, CommentArray } from 'comment-json'
 import writeFileAtomic from 'write-file-atomic'
 import { withFileLock } from './files'
 import { debug } from './logger'
+import {
+  collectVariantPlatforms,
+  extractLastPlatforms,
+  GENERATION_MARKER_PREFIX,
+  hasGenerationMarker,
+  isGenerationMarker,
+  platformsExcluding,
+  resolvePlatformUnion,
+  sortVariants,
+  upsertCurrentVariant,
+} from './platform-variants'
+import { mergeTabBarProps, renderTabBarListLines, renderTabBarPropLines } from './tab-bar'
 
 /**
  * pages.json 的读取、合并、写回
  *
- * 这个文件负责 pages.json 的全部细节：
- * - 多个平台共用一个文件时，用 #ifdef 注释区分各自的条目
+ * 这个文件负责 pages.json 的主流程：
+ * - 多个平台共用一个文件时，用 #ifdef 注释区分各自的条目（合并规则
+ *   的核心在 platform-variants.ts，数组侧的适配在这里，对象侧
+ *   tabBar 属性的适配在 tab-bar.ts）
  * - 合并时保持首页排最前
  * - 写入"由插件生成"的标记，靠它识别哪些条目是插件写的
  * - 序列化格式（缩进、换行）
@@ -357,20 +372,6 @@ function ensureHomePageFirst(pagesArray: InternalPages | undefined, homePath: st
   }
 }
 
-const GENERATION_MARKER_PREFIX = 'GENERATED BY UNI-PAGES, PLATFORM:'
-
-/** 判断注释令牌是否为 pages.json 的生成标记行 */
-function isGenerationMarker(token: CommentLineToken | { type: string, value?: string }): boolean {
-  return token.type !== 'BlankLine' && typeof token.value === 'string' && token.value.trim().startsWith(GENERATION_MARKER_PREFIX)
-}
-
-/** 判断一个 pages 数组是否由插件写入（携带生成标记） */
-function hasGenerationMarker(src: CommentArray<CommentObject> | undefined): boolean {
-  if (!src)
-    return false
-  return (src[Symbol.for('before:0') as CommentSymbol] || []).some(isGenerationMarker)
-}
-
 /**
  * 判断输出是不是带注释的多行格式（只有多行才放得下按平台区分的
  * `#ifdef` 块）。`indent: 0` 或空字符串会让 comment-json 输出成
@@ -432,414 +433,6 @@ function lineComment(value: string): CommentLineToken {
   }
 }
 
-/** 从现有 pages.json 文本里提取出来的一条 tabBar 属性 */
-interface RawTabBarProp {
-  key: string
-  /** 属性值（comment-json 解析结果，保留内部注释），用于比较和重新序列化 */
-  value: unknown
-  /** 紧挨着的第一条条件编译指令（如 `#ifdef H5 || MP-WEIXIN`、`#ifndef MP-ALIPAY`），没有包裹时为 null */
-  condition: string | null
-  /** 键前面 / 值后面的注释内容，重新原样输出 #ifndef 属性时要用 */
-  beforeComments: string[]
-  afterComments: string[]
-}
-
-/**
- * 跳过空白和注释往前走。遇到行注释或块注释时，注释里的文字（去掉
- * 首尾空白）会收集到 `comments` 里（如果传了）
- */
-function skipJsoncFiller(content: string, index: number, comments?: string[]): number {
-  let i = index
-  while (i < content.length) {
-    const ch = content[i]
-    if (ch === ' ' || ch === '\t' || ch === '\r' || ch === '\n') {
-      i++
-      continue
-    }
-    if (ch === '/' && content[i + 1] === '/') {
-      const end = content.indexOf('\n', i)
-      const stop = end === -1 ? content.length : end
-      comments?.push(content.slice(i + 2, stop).trim())
-      i = stop
-      continue
-    }
-    if (ch === '/' && content[i + 1] === '*') {
-      const end = content.indexOf('*/', i + 2)
-      const stop = end === -1 ? content.length : end + 2
-      comments?.push(content.slice(i + 2, end === -1 ? content.length : end).trim())
-      i = stop
-      continue
-    }
-    return i
-  }
-  return i
-}
-
-/** 从起始引号扫描 JSON 字符串字面量；返回闭合引号之后的下标 */
-function scanJsonString(content: string, start: number): number {
-  let i = start + 1
-  while (i < content.length) {
-    if (content[i] === '\\') {
-      i += 2
-      continue
-    }
-    if (content[i] === '"')
-      return i + 1
-    i++
-  }
-  return content.length
-}
-
-/** 扫描 `start` 处（定位在值的首字符）的 JSON 值；返回其后的下标 */
-function scanJsonValue(content: string, start: number): number {
-  const ch = content[start]
-  if (ch === '"')
-    return scanJsonString(content, start)
-  if (ch === '{' || ch === '[') {
-    let depth = 0
-    let i = start
-    while (i < content.length) {
-      const c = content[i]
-      if (c === '"') {
-        i = scanJsonString(content, i)
-        continue
-      }
-      if (c === '/' && (content[i + 1] === '/' || content[i + 1] === '*')) {
-        i = skipJsoncFiller(content, i)
-        continue
-      }
-      if (c === '{' || c === '[') {
-        depth++
-        i++
-        continue
-      }
-      if (c === '}' || c === ']') {
-        depth--
-        i++
-        if (depth === 0)
-          return i
-        continue
-      }
-      i++
-    }
-    return content.length
-  }
-  // 没带引号的字面量（数字 / true / false / null）：读到下一个分隔符为止
-  let i = start
-  while (i < content.length && !',{}[] \t\r\n"'.includes(content[i]))
-    i++
-  return i
-}
-
-/** 读取 `start` 处（其起始引号）的 JSON 字符串字面量；返回解析文本与其后的下标 */
-function readJsonStringToken(content: string, start: number): [string, number] | null {
-  const end = scanJsonString(content, start)
-  try {
-    return [JSON.parse(content.slice(start, end)) as string, end]
-  }
-  catch {
-    return null
-  }
-}
-
-/**
- * 按文本扫描 pages.json 里的 tabBar 对象，把每个顶层属性连同它的
- * 条件编译包裹一起收集起来。
- *
- * 为什么要按文本扫：comment-json 解析时会按普通 JS 对象的规则，
- * 同一个键出现多次只留最后一个，而"同一个属性按平台写多个值"靠的
- * 就是重复键。
- * 这些信息一旦解析就丢了，所以必须从原始文本里捞回来。
- * tabBar 不是对象、或文件结构看不懂时返回 undefined。
- */
-function extractTabBarProps(content: string): Map<string, RawTabBarProp[]> | undefined {
-  let i = skipJsoncFiller(content, 0)
-  if (content[i] !== '{')
-    return undefined
-  i = skipJsoncFiller(content, i + 1)
-
-  while (i < content.length) {
-    if (content[i] === '}' || content[i] !== '"')
-      return undefined
-    const token = readJsonStringToken(content, i)
-    if (!token)
-      return undefined
-    const [key, keyEnd] = token
-    i = skipJsoncFiller(content, keyEnd)
-    if (content[i] !== ':')
-      return undefined
-    i = skipJsoncFiller(content, i + 1)
-    if (key === 'tabBar') {
-      if (content[i] !== '{')
-        return undefined
-      return parseTabBarProps(content, i + 1)
-    }
-    i = skipJsoncFiller(content, scanJsonValue(content, i))
-    if (content[i] === ',')
-      i = skipJsoncFiller(content, i + 1)
-    else if (content[i] !== '}')
-      return undefined
-  }
-  return undefined
-}
-
-/** 从 tabBar 对象起始 `{` 之后开始解析其内部 */
-function parseTabBarProps(content: string, interiorStart: number): Map<string, RawTabBarProp[]> | undefined {
-  const props = new Map<string, RawTabBarProp[]>()
-  let beforeComments: string[] = []
-  let i = skipJsoncFiller(content, interiorStart, beforeComments)
-  // 条件栈：手写的条件块可以一次包住好几个属性（#ifdef 后面跟几行
-  // 属性、最后才 #endif）。栈顶就是当前属性待在哪个条件里。旧代码只
-  // 看每个属性自己紧挨着的注释，块里第二个之后的属性全部当成没有条
-  // 件处理，被别的平台顶掉或漏出去。每个处于条件中的属性注册时都
-  // 带上完整的包裹（开头指令 + 结尾 #endif），互相独立：渲染按键分
-  // 组、按平台排序会把属性的顺序打乱，谁都不依赖邻居，输出才不会
-  // 出现没闭合的 #ifndef
-  const openConditions: string[] = []
-  const applyCommentsToStack = (comments: string[]): void => {
-    for (const comment of comments) {
-      const trimmed = comment.trim()
-      if (trimmed.startsWith('#endif')) {
-        openConditions.pop()
-      }
-      else if (trimmed.startsWith('#ifdef') || trimmed.startsWith('#ifndef')) {
-        openConditions.push(trimmed)
-      }
-    }
-  }
-  applyCommentsToStack(beforeComments)
-
-  while (true) {
-    if (i >= content.length)
-      return undefined
-    if (content[i] === '}')
-      return props
-    if (content[i] === ',') {
-      const postComma: string[] = []
-      i = skipJsoncFiller(content, i + 1, postComma)
-      // 逗号后面的注释在文本上分属两处：开头的 #endif 关掉上一个属性
-      // 待着的条件（comment-json 把 #endif 输出在值的逗号之后），剩下
-      // 的跟着下一个属性走。#endif 不再挂到上一个属性身上：需要闭合
-      // 符的 #ifndef 属性在注册时已经自带了
-      applyCommentsToStack(postComma)
-      beforeComments = postComma
-      continue
-    }
-    if (content[i] !== '"')
-      return undefined
-    const token = readJsonStringToken(content, i)
-    if (!token)
-      return undefined
-    const [key, keyEnd] = token
-    i = skipJsoncFiller(content, keyEnd)
-    if (content[i] !== ':')
-      return undefined
-    i = skipJsoncFiller(content, i + 1)
-    const valueEnd = scanJsonValue(content, i)
-    const rawValue = content.slice(i, valueEnd)
-
-    const afterComments: string[] = []
-    i = skipJsoncFiller(content, valueEnd, afterComments)
-    if (content[i] !== ',' && content[i] !== '}')
-      return undefined
-
-    // 条件在应用值后面的注释之前读：那些注释（可能有关闭符或新的
-    // 指令）在文本上位于这个属性之后，属于下一个属性的世界
-    const condition = openConditions[openConditions.length - 1] ?? null
-    applyCommentsToStack(afterComments)
-
-    try {
-      const value = cjParse(rawValue)
-      let storedBefore = [...beforeComments]
-      let storedAfter = afterComments
-      if (condition?.startsWith('#ifndef')) {
-        // #ifndef 属性自带完整包裹：开头指令统一重写为本属性的条件
-        // （块里的后续属性自己的注释里没有它），结尾统一用合成的
-        // #endif。注释里其他属性留下的结构性指令（上一个块的 #endif、
-        // 重复的 #ifndef）全部滤掉——每个属性的包裹都是独立配对的，
-        // 多余的指令只会输出成悬空的空块或没闭合的块；用户写的普通
-        // 注释（说明文字）保留在指令后面
-        storedBefore = [
-          condition,
-          ...beforeComments.filter(c => !/^#(?:ifdef|ifndef|endif)/.test(c.trim())),
-        ]
-        storedAfter = ['#endif', ...afterComments.filter(c => !c.trim().startsWith('#endif'))]
-      }
-      const raw: RawTabBarProp = { key, value, condition, beforeComments: storedBefore, afterComments: storedAfter }
-      const variants = props.get(key) || []
-      variants.push(raw)
-      props.set(key, variants)
-    }
-    catch {
-      // 值解析不出来：丢弃这条，本次运行会用自己的配置重写这个属性
-    }
-  }
-}
-
-/** 一条已经分好平台的 tabBar 属性变体（`list` 以外的所有属性） */
-interface TabBarPropVariant {
-  /** 用来比较的字符串形式：值去掉注释后 JSON.stringify */
-  valueStr: string
-  /** 原始的 comment-json 值（保留内部注释），重新序列化时用 */
-  value: unknown
-  /** 拥有这个值的平台；null 表示这是手写的 `#ifndef` 属性，原样保留 */
-  platforms: string[] | null
-  /** 原样保留时需要的包裹注释 */
-  passthrough: { before: string[], after: string[] } | null
-}
-
-/**
- * 跨平台合并 tabBar 的外观属性（`list` 以外的所有属性）。
- * 规则和数组的 {@link mergePlatformItems} 一致：
- * - 包在 #ifdef 里的属性，属于注释里列出的那些平台
- * - 没包裹的属性，属于除当前平台外的所有平台（当前平台会写自己的）
- * - 一个值最后一个平台都不剩了，就删掉它
- * - 手写的 #ifndef 属性原样保留、永不删除，和数组侧的处理一致
- */
-function mergeTabBarProps(
-  existingContent: string,
-  currentProps: Record<string, unknown>,
-  contributesProps: boolean,
-  currentPlatform: string,
-  platformUnion: string[],
-): Map<string, TabBarPropVariant[]> {
-  const rawProps = extractTabBarProps(existingContent)
-  const merged = new Map<string, TabBarPropVariant[]>()
-
-  for (const [key, variants] of rawProps || []) {
-    if (key === 'list')
-      continue
-    for (const raw of variants) {
-      const valueStr = JSON.stringify(raw.value)
-      if (valueStr === undefined)
-        continue
-      if (raw.condition?.startsWith('#ifndef')) {
-        const list = merged.get(key) || []
-        list.push({ valueStr, value: raw.value, platforms: null, passthrough: { before: raw.beforeComments, after: raw.afterComments } })
-        merged.set(key, list)
-        continue
-      }
-      let platforms: string[]
-      if (raw.condition?.startsWith('#ifdef')) {
-        platforms = platformsExcluding(raw.condition.slice('#ifdef'.length).trim(), currentPlatform)
-      }
-      else {
-        platforms = platformUnion.filter(p => p !== currentPlatform)
-      }
-      if (platforms.length === 0)
-        continue
-      const list = merged.get(key) || []
-      list.push({ valueStr, value: raw.value, platforms, passthrough: null })
-      merged.set(key, list)
-    }
-  }
-
-  if (contributesProps) {
-    for (const [key, value] of Object.entries(currentProps)) {
-      const valueStr = JSON.stringify(value)
-      if (valueStr === undefined)
-        continue
-      const list = merged.get(key) || []
-      // 值相同的 #ifndef 属性优先：如果当前平台没被 #ifndef 排除，
-      // 那条手写属性本来就已经覆盖了它；如果当前平台正好被排除，
-      // 尊重手写的隐藏规则。两种情况都不该再单独写一份
-      if (list.some(v => v.platforms === null && v.valueStr === valueStr))
-        continue
-      const equal = list.find((v): v is TabBarPropVariant & { platforms: string[] } => v.platforms !== null && v.valueStr === valueStr)
-      if (equal) {
-        equal.platforms = [...equal.platforms, currentPlatform].sort()
-      }
-      else {
-        list.push({ valueStr, value, platforms: [currentPlatform], passthrough: null })
-        merged.set(key, list)
-      }
-    }
-  }
-
-  // 和 mergePlatformItems 一样按平台排好序再输出：当前平台原来的属性
-  // 变体是从文件里读出来的，读进来时会被丢掉、再由本次运行补到列表
-  // 末尾。两个平台轮流跑时，顺序就会来回跳，文件每次都有无意义的
-  // 变动。排好序之后不管谁先谁后，写出来的都一样。#ifndef 属性
-  // （platforms 为 null）排最前，这样文档承诺的"当前平台自己的值
-  // 排在后面、优先生效"在与 #ifndef 重复键的场景下依然成立
-  for (const variants of merged.values()) {
-    variants.sort((a, b) => {
-      const keyA = a.platforms === null ? '' : a.platforms.join(' || ')
-      const keyB = b.platforms === null ? '' : b.platforms.join(' || ')
-      return keyA < keyB ? -1 : keyA > keyB ? 1 : 0
-    })
-  }
-
-  return merged
-}
-
-/** 一层缩进是多少（空格数或字符串） */
-function indentUnitOf(format: PagesJsonFormatOptions | undefined): string {
-  const indent = format?.indent ?? 2
-  return typeof indent === 'number' ? ' '.repeat(indent) : indent
-}
-
-/** 给除第一行以外的每一行加缩进（用于多行值拼进属性行的场景） */
-function indentContinuation(text: string, extra: string): string {
-  return text.split('\n').map((line, index) => index === 0 ? line : extra + line).join('\n')
-}
-
-/**
- * 把 tabBar 外观属性渲染成带缩进的输出行。所有平台都适用的值直接
- * 输出；其余每个值都写在自己的 `#ifdef` 块里（同一个键出现多次），
- * uni-app 按平台剥掉注释后，每个平台只会看到一个值。
- */
-function renderTabBarPropLines(merged: Map<string, TabBarPropVariant[]>, platformUnion: string[], format: PagesJsonFormatOptions | undefined): string[] {
-  const unit = indentUnitOf(format)
-  const unionStr = platformUnion.join(' || ')
-  const lines: string[] = []
-
-  for (const [key, variants] of merged) {
-    for (const variant of variants) {
-      const valueText = indentContinuation(cjStringify(variant.value, null, format?.indent ?? 2), unit + unit)
-      const propLine = `${unit}${unit}${JSON.stringify(key)}: ${valueText},`
-      if (variant.passthrough) {
-        for (const comment of variant.passthrough.before) {
-          if (comment)
-            lines.push(`${unit}${unit}// ${comment}`)
-        }
-        lines.push(propLine)
-        for (const comment of variant.passthrough.after) {
-          if (comment)
-            lines.push(`${unit}${unit}// ${comment}`)
-        }
-        continue
-      }
-      const platformStr = (variant.platforms as string[]).join(' || ')
-      if (platformStr === unionStr) {
-        lines.push(propLine)
-      }
-      else {
-        lines.push(`${unit}${unit}// #ifdef ${platformStr}`)
-        lines.push(propLine)
-        lines.push(`${unit}${unit}// #endif`)
-      }
-    }
-  }
-  return lines
-}
-
-/** 把合并后的 tabBar list 渲染成带缩进的输出行（list 总是 tabBar 的最后一个属性） */
-function renderTabBarListLines(list: CommentArray<CommentObject>, format: PagesJsonFormatOptions | undefined): string[] {
-  const unit = indentUnitOf(format)
-  const text = cjStringify(list, null, format?.indent ?? 2)
-  return text.split('\n').map((line, index) => {
-    if (index === 0)
-      return `${unit}${unit}"list": ${line}`
-    return line === '' ? line : unit + unit + line
-  })
-}
-
-/** 把 `||` 分隔的平台列表拆开，去掉当前平台，剩下的排好序 */
-function platformsExcluding(platformList: string, currentPlatform: string): string[] {
-  return platformList.split('||').map(p => p.trim()).filter(p => p !== currentPlatform).sort()
-}
-
 /**
  * 把条目转成字符串用于比较，比较前要先把内部的 `type` 标记
  * （'home' | 'page'）拿掉。
@@ -878,12 +471,8 @@ function homeStatusCompatible(a: object, b: object): boolean {
   return typeA === undefined || typeB === undefined || typeA === typeB
 }
 
-/** 一条合并后的条目及其出现的平台 */
-interface MultiPlatformItem<T extends object> {
-  item: T
-  itemStr: string
-  platforms: string[]
-  platformStr: string
+/** 一条合并后的条目及其出现的平台（数组侧的变体形状） */
+interface MultiPlatformItem<T extends object> extends PlatformVariant<T> {
   /**
    * 手写 `#ifndef` 块原样保留所需的注释。插件只会记录"这个条目属于
    * 哪些平台"这种正向列表，表达不了"除了某平台之外"这种否定条件，
@@ -892,71 +481,6 @@ interface MultiPlatformItem<T extends object> {
    * 它紧跟的那一条。普通条目没有这个字段。
    */
   passthrough?: { before: CommentToken[], after: CommentToken[] }
-}
-
-/**
- * 从 before:0 的生成标记里读出上一次运行记录了哪些平台，
- * 去掉当前平台后返回
- */
-function extractLastPlatforms(src: CommentArray<CommentObject>, currentPlatform: string): string[] {
-  let lastPlatforms: string[] = []
-  for (const comment of (src[Symbol.for('before:0') as CommentSymbol] || [])) {
-    // comment-json v5 遇到源文本里的空行会给出没有 `value` 的
-    // BlankLine 令牌，比如手工整理过格式的 pages.json
-    if (comment.type === 'BlankLine')
-      continue
-
-    const trimmed = comment.value.trim()
-    if (trimmed.startsWith(GENERATION_MARKER_PREFIX)) {
-      lastPlatforms = platformsExcluding(trimmed.split(':')[1], currentPlatform)
-    }
-  }
-  return lastPlatforms
-}
-
-/**
- * 把所有条目记录的平台合在一起，算出"平台全集"
- *
- * 平台全集会写进生成标记，也决定哪些条目可以不加 #ifdef 直接输出：
- * 没被包裹的条目在 uni-app 条件编译下所有平台都看得见，所以只有
- * 覆盖了全部平台的条目才能"裸着"输出。
- *
- * 当前平台一定在全集里，哪怕这次运行一个条目都没贡献——比如页面
- * 只在 H5 存在，而 MP-WEIXIN 运行把每个页面都 definePage(null) 掉了。
- * 如果只从剩下的条目推全集，全集就只有 'H5'，H5 专属条目看起来
- * "覆盖了全集"，会不加包裹地输出，泄漏进微信的构建产物。老版本按
- * "最常用平台组合当默认值"的做法也有同样的问题：可能算出单平台
- * 组合，把那个平台的条目裸着输出，变成其他平台里的重复路由。这个
- * 全集同样管着 tabBar 属性的包裹：没有它，一次没有 tabBar 的运行
- * 会以为"拥有平台 == 全集"，把别的平台的 tabBar 裸着输出出去。
- *
- * 上一次运行记在标记里的平台也进全集。否则当某平台的条目都被
- * 重新扫描替换后，它"来过"这件事只记在标记里：只看条目推全集，
- * 全集会一会儿大一会儿小，条目一会儿包一会儿裸。把标记也算进来，
- * 全集就只增不减；多算的平台最多带来多余的包裹，不会泄漏。
- *
- * 只增不减的代价：彻底不用的平台会一直留在标记里。这无害，但别
- * 手工去改标记行——想清理就删掉整个生成的 pages.json，下次运行会
- * 从头生成。
- */
-/**
- * 收集所有条目记录过的平台（不算当前平台和标记里记的平台），这个
- * 集合同时也是"拥有至少一个条目"的平台集合
- */
-function collectVariantPlatforms<T extends object>(mergedMap: Map<string, MultiPlatformItem<T>[]>): Set<string> {
-  const platforms = new Set<string>()
-  for (const variants of mergedMap.values()) {
-    for (const { platforms: variantPlatforms } of variants) {
-      for (const platform of variantPlatforms)
-        platforms.add(platform)
-    }
-  }
-  return platforms
-}
-
-function resolvePlatformUnion<T extends object>(mergedMap: Map<string, MultiPlatformItem<T>[]>, currentPlatform: string, lastPlatforms: string[]): string[] {
-  const union = new Set<string>([currentPlatform, ...lastPlatforms, ...collectVariantPlatforms(mergedMap)])
-  return [...union].sort()
 }
 
 /** {@link mergePlatformItems} 的返回结果：合并后的条目和相关的平台集合 */
@@ -974,10 +498,11 @@ interface MergedPlatformItems {
 }
 
 /**
- * 合并多平台的配置条目。
+ * 合并多平台的配置条目（数组侧适配器）。
  *
  * 处理条件编译注释（#ifdef / #ifndef / #endif），把 pages.json 里
- * 已有的条目和本次扫描的条目合并成一个数组：
+ * 已有的条目和本次扫描的条目合并成一个数组；合并规则（同值叠平台、
+ * 排除当前平台、排序、并集）用 platform-variants.ts 的共享核心：
  * - 内容相同的条目自动合并，平台标识叠在一起
  * - 内容不同的条目各写各的 #ifdef 块
  * - 手写的 `#ifndef` 块原样保留：否定条件翻译不成正向平台列表，
@@ -998,6 +523,12 @@ function mergePlatformItems<T extends object = Record<string, unknown>>(source: 
 
   // 2. 遍历现有数组，逐个判断，以 uniqueKey 的值为键放进 mergedMap
   const mergedMap = new Map<string, MultiPlatformItem<T>[]>()
+  // 取这个键的变体列表，没有就建一个并注册进 mergedMap
+  const variantsOf = (uniqueKey: string): MultiPlatformItem<T>[] => {
+    const list = mergedMap.get(uniqueKey) || []
+    mergedMap.set(uniqueKey, list)
+    return list
+  }
 
   for (let i = 0; i < src.length; i++) {
     const item = src[i] as unknown as T
@@ -1020,18 +551,16 @@ function mergePlatformItems<T extends object = Record<string, unknown>>(source: 
     // 块一定出自用户之手。#endif 在输出时补写；挂在后续条目前面的
     // 游离 #endif 被上面的匹配规则跳过，重建包裹层时自然丢掉
     if (conditionalComment?.value.trim().startsWith('#ifndef')) {
-      const existing = mergedMap.get(uniqueKey) || []
-      existing.push({
-        item,
-        itemStr: stringifyForCompare(item),
-        platforms: [],
+      variantsOf(uniqueKey).push({
+        payload: item,
+        valueStr: stringifyForCompare(item),
+        platforms: null,
         platformStr: '',
         passthrough: {
           before: [...(beforeComments || [])].filter(c => !isGenerationMarker(c)),
           after: [lineComment(' #endif')],
         },
       })
-      mergedMap.set(uniqueKey, existing)
       continue
     }
 
@@ -1055,9 +584,7 @@ function mergePlatformItems<T extends object = Record<string, unknown>>(source: 
       continue
     }
 
-    const existing = mergedMap.get(uniqueKey) || []
-    existing.push({ item, itemStr: stringifyForCompare(item), platforms, platformStr: platforms.join(' || ') })
-    mergedMap.set(uniqueKey, existing)
+    variantsOf(uniqueKey).push({ payload: item, valueStr: stringifyForCompare(item), platforms, platformStr: platforms.join(' || ') })
   }
 
   // 3. 把本次扫描的条目合并进 mergedMap
@@ -1071,56 +598,21 @@ function mergePlatformItems<T extends object = Record<string, unknown>>(source: 
       continue
     }
 
-    if (!mergedMap.has(uniqueKey)) {
-      // 还没有这个条目：直接加入
-      mergedMap.set(uniqueKey, [{
-        item,
-        itemStr: stringifyForCompare(item),
-        platforms: [currentPlatform],
-        platformStr: currentPlatform,
-      }])
-      continue
-    }
-
-    // 已有同键条目：看看内容是否相同
-    const existing = mergedMap.get(uniqueKey)!
-
-    const itemStr = stringifyForCompare(item)
-    // 内容相同的手写 #ifndef 条目优先：当前平台没被排除时，那条
-    // #ifndef 已经覆盖了它；当前平台正好被排除时，隐藏扫描条目正是
-    // 尊重手写的排除规则。两种情况下再单独写一份都会产生重复路由
-    // 或覆盖用户的条件
-    if (existing.some(val => val.passthrough && val.itemStr === itemStr && homeStatusCompatible(val.item, item)))
-      continue
-    const equalObj = existing.find(val => !val.passthrough && val.itemStr === itemStr && homeStatusCompatible(val.item, item))
-    if (equalObj) {
-      equalObj.platforms.push(currentPlatform)
-      equalObj.platforms.sort()
-      equalObj.platformStr = equalObj.platforms.join(' || ')
-    }
-    else {
-      existing.push({
-        item,
-        itemStr,
-        platforms: [currentPlatform],
-        platformStr: currentPlatform,
-      })
-    }
+    // 规则见 platform-variants 的 upsertCurrentVariant：同值叠平台、
+    // 同值 #ifndef 优先；compatible 用来区分两个平台各自声明的不同
+    // 首页（homeStatusCompatible）
+    upsertCurrentVariant(variantsOf(uniqueKey), item, stringifyForCompare(item), currentPlatform, homeStatusCompatible)
   }
 
   // 4. 遍历 mergedMap 生成结果：CommentArray<CommentObject>
   const result = new CommentArray<CommentObject>()
 
-  // 输出前按平台列表把每个键的变体排好序：当前平台的变体在步骤 2
-  // 里被丢掉、步骤 3 里又补到列表末尾，两个平台轮流运行（dev:h5 和
-  // dev:mp-weixin 交替写）时，同一键的变体顺序就会来回翻转——条件
-  // 编译的结果没变，但文件每次都有无意义的变动。platformStr 由排好
-  // 序的平台列表拼成（platformsExcluding / resolvePlatformUnion 都
-  // 排过序），手写 #ifndef 条目是空串、排最前，所以不管写入顺序
-  // 怎样，输出都一样。排序是稳定的：并列时（同键两组平台列表相同，
-  // 只在手写文件里出现）保持文件里的顺序
+  // 输出前按平台列表把每个键的变体排好序（规则见 platform-variants
+  // 的 sortVariants）：当前平台的变体在步骤 2 里被丢掉、步骤 3 里又
+  // 补到列表末尾，两个平台轮流运行（dev:h5 和 dev:mp-weixin 交替
+  // 写）时，同一键的变体顺序就会来回翻转
   for (const variants of mergedMap.values()) {
-    variants.sort((a, b) => (a.platformStr < b.platformStr ? -1 : a.platformStr > b.platformStr ? 1 : 0))
+    sortVariants(variants)
   }
 
   // 只有覆盖全部平台的条目才能不加 #ifdef 直接输出：没包裹的条目
@@ -1140,7 +632,7 @@ function mergePlatformItems<T extends object = Record<string, unknown>>(source: 
 
   // 按插入顺序处理元素
   for (const [_, list] of mergedMap) {
-    for (const { item, platformStr, passthrough } of list) {
+    for (const { payload: item, platformStr, passthrough } of list) {
       result.push(item as unknown as CommentObject)
 
       // 手写 #ifndef 块：把原始包裹注释原样放回去
