+++ packages/core/src/platform-variants.ts
@@ -1,0 +1,187 @@
+import type { CommentArray, CommentLineToken, CommentObject, CommentSymbol } from 'comment-json'
+
+/**
+ * 平台变体表：跨平台合并的核心规则
+ *
+ * pages.json 是所有平台共用的一个文件，「同一份内容在不同平台各有一
+ * 个变体」靠 #ifdef 条件编译注释区分。一个变体 = 载荷 + 拥有平台；
+ * 手写的 #ifndef 变体没有正向平台列表，按「原样保留」处理。
+ *
+ * 合并规则只在这里写一遍，两类调用方共用：
+ * 1. 排除当前平台：现有变体的平台列表先去掉当前平台——本次运行总会
+ *    写出自己的版本（{@link platformsExcluding}）
+ * 2. 内容相同则叠加平台：和现有变体内容相同的本次载荷不另开变体，
+ *    把当前平台叠进它的平台列表；内容相同的手写 #ifndef 变体优先，
+ *    跳过不写（{@link upsertCurrentVariant}）
+ * 3. 手写 #ifndef 永不修改：否定条件翻译不成正向平台列表，变体连同
+ *    原始包裹一起原样输出，也永不清理（用 platforms 为 null 表示）
+ * 4. 按 platformStr 稳定排序：两个平台轮流跑时输出顺序不来回跳，
+ *    手写 #ifndef（空 platformStr）排最前（{@link sortVariants}）
+ * 5. 覆盖平台并集才裸输出：没被包裹的条目所有平台都看得见，覆盖面
+ *    小的变体必须包进 #ifdef，免得漏进其他平台的构建产物（比较
+ *    platformStr 与并集串，在两侧的渲染处执行）
+ *
+ * 两类适配器往里喂数据：数组侧（pages / subPackages / tabBar.list）
+ * 从 comment-json 的注释符号里读 #ifdef（pages-json.ts 的
+ * mergePlatformItems），对象侧（tabBar 外观属性）从原始文本里读重复
+ * 键（tab-bar.ts 的 mergeTabBarProps——comment-json 解析时同一个键
+ * 只留最后一个，「同一个属性按平台写多个值」只能靠文本扫描捞回来）。
+ */
+
+/** 一个跨平台变体：载荷 + 拥有平台 */
+export interface PlatformVariant<P> {
+  /** 变体载荷（pages.json 条目对象 / tabBar 属性值） */
+  payload: P
+  /** 参与内容比较的字符串形式 */
+  valueStr: string
+  /**
+   * 拥有这个变体的平台，排好序；null 表示手写的 `#ifndef` 变体，
+   * 原样保留、永不修改
+   */
+  platforms: string[] | null
+  /** platforms 的 ` || ` 连接，排序与渲染用；null 时为空串 */
+  platformStr: string
+}
+
+/**
+ * 把 `||` 分隔的平台列表拆开，去掉当前平台，剩下的排好序（规则 1）
+ */
+export function platformsExcluding(platformList: string, currentPlatform: string): string[] {
+  return platformList.split('||').map(p => p.trim()).filter(p => p !== currentPlatform).sort()
+}
+
+/**
+ * 把本次运行的一个载荷并入变体列表（规则 2）：
+ * - 内容相同的手写 #ifndef 变体优先：当前平台没被排除时，那条
+ *   #ifndef 已经覆盖了它；当前平台正好被排除时，尊重手写的隐藏规
+ *   则。两种情况下再单独写一份都会产生重复路由或覆盖用户的条件
+ * - 内容相同的普通变体：当前平台叠进它的平台列表
+ * - 没有内容相同的变体：新建只属于当前平台的变体
+ *
+ * @param variants - 同一个键下已有的变体列表，原地修改
+ * @param payload - 本次运行的载荷
+ * @param valueStr - 载荷的比较字符串
+ * @param currentPlatform - 当前平台标识（大写）
+ * @param compatible - 可选的额外相容判断。数组侧用它区分两个平台各自
+ * 声明的不同首页：内容相同但首页标记冲突时不合并
+ */
+export function upsertCurrentVariant<P>(
+  variants: PlatformVariant<P>[],
+  payload: P,
+  valueStr: string,
+  currentPlatform: string,
+  compatible?: (a: P, b: P) => boolean,
+): void {
+  const isSame = (variant: PlatformVariant<P>): boolean =>
+    variant.valueStr === valueStr && (compatible ? compatible(variant.payload, payload) : true)
+
+  if (variants.some(variant => variant.platforms === null && isSame(variant)))
+    return
+
+  const equal = variants.find((variant): variant is PlatformVariant<P> & { platforms: string[] } =>
+    variant.platforms !== null && isSame(variant))
+  if (equal) {
+    equal.platforms.push(currentPlatform)
+    equal.platforms.sort()
+    equal.platformStr = equal.platforms.join(' || ')
+    return
+  }
+
+  variants.push({ payload, valueStr, platforms: [currentPlatform], platformStr: currentPlatform })
+}
+
+/**
+ * 输出前按平台列表把变体排好序（规则 4）
+ *
+ * 当前平台的变体在收集时被排除、并入时又补到列表末尾，两个平台轮流
+ * 运行（dev:h5 和 dev:mp-weixin 交替写）时，同一个键的变体顺序就会
+ * 来回翻转——条件编译的结果没变，但文件每次都有无意义的变动。
+ * platformStr 由排好序的平台列表拼成，手写 #ifndef 变体是空串、排最
+ * 前。排序是稳定的：并列时（同键两组平台列表相同，只出现在手写文件
+ * 里）保持收集顺序
+ */
+export function sortVariants<P>(variants: PlatformVariant<P>[]): void {
+  variants.sort((a, b) => (a.platformStr < b.platformStr ? -1 : a.platformStr > b.platformStr ? 1 : 0))
+}
+
+/**
+ * 收集所有变体记录过的平台（手写 #ifndef 不算，它没有正向列表），
+ * 这个集合同时也是「拥有至少一个变体」的平台集合
+ */
+export function collectVariantPlatforms<P>(mergedMap: Map<string, PlatformVariant<P>[]>): Set<string> {
+  const platforms = new Set<string>()
+  for (const variants of mergedMap.values()) {
+    for (const { platforms: variantPlatforms } of variants) {
+      if (variantPlatforms) {
+        for (const platform of variantPlatforms)
+          platforms.add(platform)
+      }
+    }
+  }
+  return platforms
+}
+
+/**
+ * 把所有变体记录的平台合在一起，算出「平台全集」
+ *
+ * 平台全集会写进生成标记，也决定哪些变体可以不加 #ifdef 直接输出：
+ * 没被包裹的条目在 uni-app 条件编译下所有平台都看得见，所以只有
+ * 覆盖了全部平台的变体才能"裸着"输出（规则 5）。
+ *
+ * 当前平台一定在全集里，哪怕这次运行一个变体都没贡献——比如页面
+ * 只在 H5 存在，而 MP-WEIXIN 运行把每个页面都 definePage(null) 掉了。
+ * 如果只从剩下的条目推全集，全集就只有 'H5'，H5 专属条目看起来
+ * "覆盖了全集"，会不加包裹地输出，泄漏进微信的构建产物。老版本按
+ * "最常用平台组合当默认值"的做法也有同样的问题：可能算出单平台
+ * 组合，把那个平台的条目裸着输出，变成其他平台里的重复路由。这个
+ * 全集同样管着 tabBar 属性的包裹：没有它，一次没有 tabBar 的运行
+ * 会以为"拥有平台 == 全集"，把别的平台的 tabBar 裸着输出出去。
+ *
+ * 上一次运行记在标记里的平台也进全集（{@link extractLastPlatforms}）。
+ * 否则当某平台的条目都被重新扫描替换后，它"来过"这件事只记在标记
+ * 里：只看条目推全集，全集会一会儿大一会儿小，条目一会儿包一会儿
+ * 裸。把标记也算进来，全集就只增不减；多算的平台最多带来多余的包
+ * 裹，不会泄漏。
+ *
+ * 只增不减的代价：彻底不用的平台会一直留在标记里。这无害，但别
+ * 手工去改标记行——想清理就删掉整个生成的 pages.json，下次运行会
+ * 从头生成。
+ */
+export function resolvePlatformUnion<P>(mergedMap: Map<string, PlatformVariant<P>[]>, currentPlatform: string, lastPlatforms: string[]): string[] {
+  const union = new Set<string>([currentPlatform, ...lastPlatforms, ...collectVariantPlatforms(mergedMap)])
+  return [...union].sort()
+}
+
+export const GENERATION_MARKER_PREFIX = 'GENERATED BY UNI-PAGES, PLATFORM:'
+
+/** 判断注释令牌是否为 pages.json 的生成标记行 */
+export function isGenerationMarker(token: CommentLineToken | { type: string, value?: string }): boolean {
+  return token.type !== 'BlankLine' && typeof token.value === 'string' && token.value.trim().startsWith(GENERATION_MARKER_PREFIX)
+}
+
+/** 判断一个 pages 数组是否由插件写入（携带生成标记） */
+export function hasGenerationMarker(src: CommentArray<CommentObject> | undefined): boolean {
+  if (!src)
+    return false
+  return (src[Symbol.for('before:0') as CommentSymbol] || []).some(isGenerationMarker)
+}
+
+/**
+ * 从 before:0 的生成标记里读出上一次运行记录了哪些平台，
+ * 去掉当前平台后返回
+ */
+export function extractLastPlatforms(src: CommentArray<CommentObject>, currentPlatform: string): string[] {
+  let lastPlatforms: string[] = []
+  for (const comment of (src[Symbol.for('before:0') as CommentSymbol] || [])) {
+    // comment-json v5 遇到源文本里的空行会给出没有 `value` 的
+    // BlankLine 令牌，比如手工整理过格式的 pages.json
+    if (comment.type === 'BlankLine')
+      continue
+
+    const trimmed = comment.value.trim()
+    if (trimmed.startsWith(GENERATION_MARKER_PREFIX)) {
+      lastPlatforms = platformsExcluding(trimmed.split(':')[1], currentPlatform)
+    }
+  }
+  return lastPlatforms
+}
